Change Stocker to avoid warning errors.
According to the lastest Pandas, .ix format is deprecated. And it will cause warnings if I follow the instructions. I change the .ix format to .loc format to avoid it.

Also, I find it really hard to install fbprophet on Windows10. But by "conda install -c conda-forge fbprophet" it can be installed easily. So I add it in README.md.